requirements.txt
customtkinter
pandas
numpy
scipy
matplotlib
llama-cpp-python
duckduckgo-search
huggingface_hub
pyinstaller
